requirements.txt
load_dotenv
flask
waitress
langchain
ctransformers==0.2.5
